requirements.txt
Flask
pymongo[srv]
Flask-PyMongo
PyJWT
gunicorn
python-dotenv